issue #180 refactor: create SharedSchedulePage, move schedule-related code's directory, and remove unused InviteUser and GroupSchedulePage component
