+++ src/pages/index.jsx
@@ -1,13 +1,13 @@
 import CommunityPage from "./CommunityPage/CommunityPage";
 import ErrorPage from "./ErrorPage/ErrorPage";
 import GroupPage from "./GroupPage/GroupPage";
-import GroupSchedulePage from "./GroupSchedulePage";
 import LandingPage from "./LandingPage/LandingPage";
 import LoginPage from "./LoginPage/LoginPage";
 import MyPage from "./MyPage/MyPage";
 import PersonalSchedulePage from "./PersonalSchedulePage/PersonalSchedulePage";
 import Root from "./Root";
 import SettingPage from "./SettingPage/SettingPage";
+import SharedSchedulePage from "./SharedSchedulePage";
 import SignUpPage from "./SignUpPage/SignUpPage";
 
 export {
@@ -17,7 +17,7 @@ export {
 	Root,
 	SignUpPage,
 	PersonalSchedulePage,
-	GroupSchedulePage,
+	SharedSchedulePage,
 	SettingPage,
 	CommunityPage,
 	GroupPage,
